feat(node+bukkit+bungeecord): Update to CloudNet 4.0.0-RC11
This patch updates the sample project to target the newly release CloudNet version 4.0.0-RC11.
In doing so, this project now requires to be built with Java SDK 23 or newer.

In addition, the Gradle samples have moved to a version catalog to ease the management and updates of dependencies and keeps all dependency versions in sync.

Due to the latest update, the RPC samples have been updated as well.