feat(cloudnet): Add new Command-Scheduler module
This change adds a new example in the form of the "Command-Scheduler" module.
This module allows administrators to schedule/queue up commands to run at specified times with a configurable delay between each command in the script.

The module shows how to use the database and document APIs of CloudNet and how to handle recurring tasks on the node.