feat(command-scheduler): Major overhaul of Command-Scheduler
This patch changes quite a bit about the Command-Scheduler module.

With the recently added `sleep` command, the need to have a command delay is gone. This also allows the entirety of the script to be incorporated into the schedule, leaving a list of commands inside that.
In addition, we now support the Unix Cron syntax for schedule expressions. With this we also simplified the code that is responsible for determining the next execution time of schedules, which comes at the cost of not being able to properly display multiple times.