fix(command-scheduler): Rework determining the next execution time
This patch attempts to correct the incorrectly determined time to the next execution.
On the hand the determination is now always based on the relative time given to the schedule and times in the past of that time are not filtered out.
The execution engine now checks for that condition itself.
The command uses the current time to attempt a properly determined next execution time. This is currently not perfectly done as human-readable expressions have no proper way restrict them to future times (unlike cron expressions).