feat(command-scheduler): Properly determine the correct execution time when multiple dates are present.
The human-readable expressions allow for more than one execution rule to be present.
This patch makes sure that the next execution time is determined correctly by sorting according to a few simple rules. In Java, comparing is always done on pairs of objects. The implemented comparison compares to execution times to a given relative time, typically the current time.

1. If one of the times is in the past and the other in the future, the one in the future is sorted first.
2. If both are equal in regard to rule 1, the time that is least distant from the current time is sorted first.

This way the next time that is least distant in the future is sorted first.